Make amounts more granular
- Offer request is now a f64 (of Eth) rather than u64
- EthereumQuantity and BitcoinQuantity now store wei and satoshis
  respectively
- Converting EthereumQuantity to a f64 of Eth is TODO

+++ trading_client/src/offer.rs
@@ -107,7 +107,7 @@ pub fn run(
     trading_api_url: TradingApiUrl,
     symbol: Symbol,
     order_type: OrderType,
-    amount: u64,
+    amount: f64,
 ) -> Result<String, TradingServiceError> {
     let offer_request_body = BuyOfferRequestBody::new(amount);
